abstracted settings layout into class + appropriate modifications
