+++ PyPaintMainUi.py
@@ -2,6 +2,7 @@ from PyQt5.QtWidgets import QMainWindow, QVBoxLayout, QHBoxLayout, QPushButton, 
 from PyQt5.QtGui import QColor
 from PyQt5.QtCore import Qt
 from PaintWidget import PaintWidget
+from Settings import Settings
 from ColorHistory import ColorHistory
 
 class PyPaintMainUi(QMainWindow):
@@ -11,7 +12,6 @@ class PyPaintMainUi(QMainWindow):
         self.move(100, 100)
 
         self._addWidgets()
-        self._linkActions()
 
     def _addWidgets(self):
         mainLayout = QVBoxLayout()
@@ -21,75 +21,10 @@ class PyPaintMainUi(QMainWindow):
 
         mainLayout.addSpacing(1)
 
-        sliderLayout = QHBoxLayout()
-
-        sliderLayout.addWidget(QLabel("Circular Cursor:"))
-        self.cursorCheckBox = QCheckBox()
-        self.cursorCheckBox.setChecked(self.paintWidget.showCursor)
-        sliderLayout.addWidget(self.cursorCheckBox)
-        sliderLayout.addSpacing(20)
-
-        self.sizeSlider = QSlider()
-        self.sizeSlider.setOrientation(Qt.Horizontal)
-        self.sizeSlider.setMinimum(1)
-        self.sizeSlider.setMaximum(50)
-        self.sizeSlider.setValue(self.paintWidget.curSize)
-
-        sliderLayout.addWidget(QLabel("Brush Size:"))
-        sliderLayout.addWidget(self.sizeSlider)
-
-        mainLayout.addLayout(sliderLayout)
-
-        mainLayout.addSpacing(1)
-
-        radiosLayout = QHBoxLayout()
-        radiosLayout.addWidget(QLabel("Brush Styles:"))
-        self.solidBrush = QRadioButton("solid")
-        self.solidBrush.setChecked(True)
-        self.sprayBrush = QRadioButton("spray")
-        self.eraseBrush = QRadioButton("erase")
-
-        radiosLayout.addWidget(self.solidBrush)
-        radiosLayout.addWidget(self.sprayBrush)
-        radiosLayout.addWidget(self.eraseBrush)
-
-        mainLayout.addLayout(radiosLayout)
-        mainLayout.addSpacing(1)
-
-
-        resizeLayout = QHBoxLayout()
-
-        resizeLayout.addWidget(QLabel("Image/Window Resize Style:"))
-
-        resizeLayout.addWidget(QLabel("Enlarge:"), 0, Qt.AlignRight)
-        self.enlargeComboBox = QComboBox()
-        self.enlargeComboBox.addItems(["extend", "scale"])
-        resizeLayout.addWidget(self.enlargeComboBox)
-
-        resizeLayout.addWidget(QLabel("Shrink:"), 0, Qt.AlignRight)
-        self.shrinkComboBox = QComboBox()
-        self.shrinkComboBox.addItems(["crop", "scale"])
-        resizeLayout.addWidget(self.shrinkComboBox)
-
-        mainLayout.addLayout(resizeLayout)
-
-        mainLayout.addSpacing(1)
-
-        buttonsLayout = QHBoxLayout()
-
-        self.clearButton = QPushButton("Clear")
-
-        self.colorButton = QPushButton("Set Color")
-
-        self.saveButton = QPushButton("Save")
-
-        buttonsLayout.addWidget(self.clearButton)
-        buttonsLayout.addWidget(self.colorButton)
-        buttonsLayout.addWidget(self.saveButton)
-
-        mainLayout.addLayout(buttonsLayout)
-
         self.colorHistory = ColorHistory(self.paintWidget)
+        self.settings = Settings(self.paintWidget, self.colorHistory)
+
+        mainLayout.addWidget(self.settings)
 
         columnLayout = QHBoxLayout()
         columnLayout.addLayout(mainLayout)
@@ -100,21 +35,5 @@ class PyPaintMainUi(QMainWindow):
         widget.setLayout(columnLayout)
 
         self.setCentralWidget(widget)
-
-    def _linkActions(self):
-        self.clearButton.clicked.connect(self.paintWidget.clearImage)
-        self.colorButton.clicked.connect(self.changeColor)
-        self.saveButton.clicked.connect(self.paintWidget.saveImage)
-        self.sizeSlider.valueChanged.connect(self.paintWidget.setBrushSize)
-        self.solidBrush.toggled.connect(self.paintWidget.setSolidBrush)
-        self.sprayBrush.toggled.connect(self.paintWidget.setSprayBrush)
-        self.eraseBrush.toggled.connect(self.paintWidget.setEraseBrush)
-        self.enlargeComboBox.activated[str].connect(self.paintWidget.setEnlargeStyle)
-        self.shrinkComboBox.activated[str].connect(self.paintWidget.setShrinkStyle)
-        self.cursorCheckBox.stateChanged.connect(self.paintWidget.setShowCursor)
-
-    def changeColor(self):
-        color = self.paintWidget.changeColor()
-        self.colorHistory.addColor(color)
